+++ mage/src/main/java/mil/nga/giat/mage/newsfeed/ObservationFeedFragment.java
@@ -284,7 +284,7 @@ public class ObservationFeedFragment extends Fragment implements IObservationEve
 			Intent intent;
 
 			// show form picker or go to
-			JsonArray formDefinitions = EventHelper.getInstance(getActivity()).getCurrentEvent().getForms();
+			JsonArray formDefinitions = EventHelper.getInstance(getActivity()).getCurrentEvent().getNonArchivedForms();
 			if (formDefinitions.size() == 0) {
 				intent = new Intent(getActivity(), ObservationEditActivity.class);
 			} else if (formDefinitions.size() == 1) {